requirements.txt
accelerate==1.2.0
aiohttp==3.11.10
backports.tarfile==1.2.0
colorama==0.4.6
future==1.0.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
librosa==0.10.2.post1
multiprocess==0.70.16
openai==1.57.0
pandas==2.2.3
pillow==11.0.0
pip-chill==1.0.3
pyarrow==18.1.0
pyaudio==0.2.14
pymemcache==4.0.0
pyperclip==1.9.0
python-dotenv==1.0.1
python-telegram-bot==21.8
speechrecognition==3.11.0
tomli==2.0.1
transformers==4.47.0
virtualenv==20.27.1
xxhash==3.5.0